Refactor: EmailRouter
