+++ package.json
@@ -12,6 +12,7 @@
   },
   "license": "GPL-3.0",
   "dependencies": {
+    "cors": "^2.8.5",
     "dotenv": "^16.0.2",
     "express": "^4.18.1",
     "helmet": "^6.0.0",
